Renamed makeReport.py to bgcval2_make_report.py
