+++ bgcval2/bgcval.py
@@ -25,7 +25,7 @@
 #####
 #
 """
-.. module:: theWholePackage
+.. module:: bgcval
    :platform: Unix
    :synopsis: A script the run the entire suite of analyses and produce an html report.
 .. moduleauthor:: Lee de Mora <ledm@pml.ac.uk>
@@ -42,7 +42,7 @@ from .bgcvaltools.downloadFromMass import downloadMass, findLastFinishedYear
 from .analysis_timeseries import analysis_timeseries, singleTimeSeries, singleTimeSeriesProfile
 from .analysis_timeseries import level1KeysDict, physKeysDict, bgcKeysDict, keymetricsfirstDict
 from .analysis_p2p import analysis_p2p, p2pDict_level2, p2pDict_physics, single_p2p
-from .makeReport import html5Maker
+from .bgcval2_make_report import html5Maker
 from .UKESMpython import folder
