+++ setup.py
@@ -199,7 +199,7 @@ setup(
             'analysis_p2p = bgcval2.analysis_p2p:run',
             'analysis_timeseries = bgcval2.analysis_timeseries:main',
             'bgcval = bgcval2.bgcval:run',
-            'makeReport = bgcval2.makeReport:main',
+            'bgcval2_make_report = bgcval2.bgcval2_make_report:main',
         ],
     },
     cmdclass={
